+++ src/index.js
@@ -5,11 +5,13 @@ import remove from './actions/remove';
 import reportError from './actions/reportError';
 import withStorageItemIdentifier from './component/withStorageItemIdentifier';
 import storageWorker from './storage/worker';
+import resolveItemIdentifier from './service/resolveItemIdentifier';
 
 const createLoadAction = ({ storeName, errors }) => load({ storeName, errors });
 const createSaveAction = ({ storeName }) => save({ storeName });
 const createRemoveAction = ({ storeName }) => remove({ storeName });
 const createReportErrorAction = ({ storeName }) => reportError({ storeName });
+const createItemIdentifierResolver = ({ storeName }) => resolveItemIdentifier({ storeName });
 const createStorageReducer = (storageConfig) => {
     storageWorker.initiate({ storeName: storageConfig.storeName });
     return withStorageReducer(storageConfig);
@@ -21,5 +23,6 @@ export {
     createRemoveAction,
     createReportErrorAction,
     createStorageReducer,
+    createItemIdentifierResolver,
     withStorageItemIdentifier,
 };